moving flatten function to utils dir
